backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-genai>=0.3.0
elevenlabs>=0.2.26
websockets>=12.0
httpx>=0.27.0
